your_project/requirements.txt
Flask
transformers
torch
pillow
